% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabyl_tpct.R
\name{tabyl_tpct}
\alias{tabyl_tpct}
\title{A Table Function}
\usage{
tabyl_tpct(x)
}
\arguments{
\item{x}{Your vector or data frame column}
}
\description{
This function creates a table with percentage formatting and totals using
functions from the janitor package
}
\examples{
tabyl_tpct()
}
\keyword{table}
